refactor: move favicons to custom
