+++ src/custom/favicons.js
